requirements.txt
google-generativeai
keyboard
pyperclip
customtkinter
packaging
pystray
Pillow
